Extract iliffe vector into its own file Save/load SME structures as compressed npz files now
